Remove hashmap
The hashmap solution had bizarre unpredictable behavior when adding
more than 5 answers, and when the answer strings were longer than
~5 characters. It seems Arduino's String library has a well-known bug
in it, relating to the freeing of String objects. There's a "one-liner"
fix that might make behavior stable...it might also be possible to
switch to char* instead of String objects, but that had its own challenges.

For now, I'm just reverting back to the previous behavior, even though I
hate it and it makes me sad.